+++ src/common/services/provideServices.ts
@@ -1,3 +1,4 @@
+import { FetchHttpClient } from '@shlinkio/shlink-js-sdk/browser';
 import { ShlinkWebComponent } from '@shlinkio/shlink-web-component';
 import type Bottle from 'bottlejs';
 import type { ConnectDecorator } from '../../container/types';
@@ -8,14 +9,13 @@ import { MainHeader } from '../MainHeader';
 import { ScrollToTop } from '../ScrollToTop';
 import { ShlinkVersionsContainer } from '../ShlinkVersionsContainer';
 import { ShlinkWebComponentContainer } from '../ShlinkWebComponentContainer';
-import { HttpClient } from './HttpClient';
 
 export const provideServices = (bottle: Bottle, connect: ConnectDecorator) => {
   // Services
   bottle.constant('window', window);
   bottle.constant('console', console);
   bottle.constant('fetch', window.fetch.bind(window));
-  bottle.service('HttpClient', HttpClient, 'fetch');
+  bottle.service('HttpClient', FetchHttpClient, 'fetch');
 
   // Components
   bottle.serviceFactory('ScrollToTop', () => ScrollToTop);