+++ test/api/services/ShlinkApiClientBuilder.test.ts
@@ -12,29 +12,26 @@ describe('ShlinkApiClientBuilder', () => {
 
   it('creates new instances when provided params are different', async () => {
     const builder = createBuilder();
-    const [firstApiClient, secondApiClient, thirdApiClient] = await Promise.all([
-      builder(server({ url: 'foo', apiKey: 'bar' })),
-      builder(server({ url: 'bar', apiKey: 'bar' })),
-      builder(server({ url: 'bar', apiKey: 'foo' })),
-    ]);
+    const firstApiClient = builder(server({ url: 'foo', apiKey: 'bar' }));
+    const secondApiClient = builder(server({ url: 'bar', apiKey: 'bar' }));
+    const thirdApiClient = builder(server({ url: 'bar', apiKey: 'foo' }));
 
-    expect(firstApiClient).not.toBe(secondApiClient);
-    expect(firstApiClient).not.toBe(thirdApiClient);
-    expect(secondApiClient).not.toBe(thirdApiClient);
+    expect(firstApiClient === secondApiClient).toEqual(false);
+    expect(firstApiClient === thirdApiClient).toEqual(false);
+    expect(secondApiClient === thirdApiClient).toEqual(false);
   });
 
-  it('returns existing instances when provided params are the same', async () => {
+  it('returns existing instances when provided params are the same', () => {
     const builder = createBuilder();
     const selectedServer = server({ url: 'foo', apiKey: 'bar' });
-    const [firstApiClient, secondApiClient, thirdApiClient] = await Promise.all([
-      builder(selectedServer),
-      builder(selectedServer),
-      builder(selectedServer),
-    ]);
 
-    expect(firstApiClient).toBe(secondApiClient);
-    expect(firstApiClient).toBe(thirdApiClient);
-    expect(secondApiClient).toBe(thirdApiClient);
+    const firstApiClient = builder(selectedServer);
+    const secondApiClient = builder(selectedServer);
+    const thirdApiClient = builder(selectedServer);
+
+    expect(firstApiClient === secondApiClient).toEqual(true);
+    expect(firstApiClient === thirdApiClient).toEqual(true);
+    expect(secondApiClient === thirdApiClient).toEqual(true);
   });
 
   it('does not fetch from state when provided param is already selected server', () => {
@@ -42,7 +39,7 @@ describe('ShlinkApiClientBuilder', () => {
     const apiKey = 'apiKey';
     const apiClient = buildShlinkApiClient(fromPartial({}))(server({ url, apiKey }));
 
-    expect(apiClient['baseUrl']).toEqual(url); // eslint-disable-line @typescript-eslint/dot-notation
-    expect(apiClient['apiKey']).toEqual(apiKey); // eslint-disable-line @typescript-eslint/dot-notation
+    expect(apiClient['serverInfo'].baseUrl).toEqual(url); // eslint-disable-line @typescript-eslint/dot-notation
+    expect(apiClient['serverInfo'].apiKey).toEqual(apiKey); // eslint-disable-line @typescript-eslint/dot-notation
   });
 });